tts: use static, not dynamic total time early on
There are currently two estimates for the total reading time:
  static  -- based on average word length and words per minute;
  dynamic -- based on the current progress through the document.

The static value gives a reasonable estimate (within approx.
+/- 0.05 of the actual time).

The dynamic value provides a chaotic estimate early on, but
stabilises to a better estimate than the static value.

Unscientific testing (including on very large documents) suggests
that at 0.1% complete the dynamic value is stable enough to be
used.