build: perform a full clean when building debian packages
There are two issues that this solves:

  1.  Running `make check` on the Ubuntu PPA build machines fails.

This is because of issues with using the auto-generated files
from different versions of autoconf/automake (specifically the
oneiric versions). Running `git clean -fxd` fixes this.

  2.  Running `dput` fails if there are older source packages.

This is because dput attempts to re-push these as a side-affect
of the wildcards used by build.sh. Instead of extracting the
version from debian/changelog, simply run the doclean target.