xml: move parser state into an internal data structure
This is moving the xml reader to the point where all its parser
state is tracked in a separate structure with the ability to
save/restore that state.

This is used internally for the xmlns attribute parsing and will
be used by the handling of empty elements containing attributes.

This then supports exposing the save/restore functionality to
consumers. The primary consumer for this will be the RDF/XML
parser that needs multi-pass attribute scanning for rdf:about,
rdf:ID, etc. and metadata attributes like dc:publisher.