xmlreader: remove comparison with uri objects
Now that the xmlreader API supports contexts, there is no need to
support comparison with uri objects -- that is an import from the
older parser logic using libxml and rdf::uri objects.

The only place where this is still used is in the mimetype code
for xml:lang. This cannot currently use a context due to object
initialization order, so compare the namespaceUri and nodeName
properties explicitly. NOTE: The mimetype information should be
loaded lazily.