mimetypes: move database entry lookups to the match calls.
The debian build is failing when running the language tests, as
it is not loading the gzip mimetype data in time for the test run.
This is a symptom of the mimetype objects and the mimetype
database being sensitive to initialization order.

This change moves the point at which the mimetype object looks up
the mimetype entry to the call to match. This then allows the
mimetype database to be lazily loaded when this call is made,
which will be done in a future change.