data/languages-extra.dat: Don't associate language tags with their primary ISO 639-5 code.
The main issue here is that ISO 639-5 is not fine grained enough to
accurately associate the language code with it's language family.
The ISO 639-6 codes which do are not specified in the IANA Sub-Tag
Repository, are not AFAICS officially standardized and do not have
an easily accessible/processible data source so cannot be used.