tts/text_reader.cpp: report Lm, Mc, Me and Mn characters as errrors.
The Lm, Mc, Me and Mn general categories are used for combining
characters such as accents. At the moment the tokenizer does not
properly handle these characters, so instead of treating them as
whitespace characters (like the C and Z classes) this change
causes them to be reported as error tokens.

This allows the characters to be investigated in the token stream
and makes it clear that there is a deficiency in the tokenizer's
state machine.

NOTE: The error token is used here to represent an internal error,
not a text encoding error (e.g. using a Unicode codepoint outside
the valid Unicode range).