tts/word_stream.cpp: limit the text_event queue to one text_reader item
The idea here is to keep the text_event queue as small as possible
to reduce the memory overhead and to increase performance, while
still allowing more complex event processing (e.g. for numbers)
that result in one or more text events.

In the future, the scanner can read a clause at a time to allow
for better context identification (ordinal numbers, dates, times,
etc.) while still keeping the overhead low.