data/locale: express the rules for English cardinal numbers in external files
This is the initial draft of the rules for cardinal numbers in
English using both the long scale and short scale representation.

These files are not currently used by the text-to-speech engine,
but the intention is to:

  1.  pass the document locale to the word_stream constructor;
  2.  load the correct cardinal data file(s) for that locale;
  3.  read the data into a modified number_words data structure;
  4.  use the number_words object in parse_cardinal_number to
      generate the correct word sequences.

These will still have tests in tests/parse-text/cardinal-numbers
to ensure that the number processor and data is correct.