espeak: fix incorrect word offset
Observed Behviour: When using cainteoir-gtk, the total time display
sometimes displayed 00:00:00.0 instead of the actual (estimated)
total.

This is because tts::speech::totalTime() and tts::speech::completed()
could return very large values.

This in turn was because tts::speech::onspeaking(pos, len) could be
passed a value of |pos=-1| which then caused speech_impl's mProgress,
mCompleted and mTotalTime to be very large.

This originated in the espeak_tts_callback handler. Examination of
the code revealed that pos was being passed:

	event->text_position - 1

which can be evaluated as |-1| when |event->text_position| is 0.
This was introduced in commit 0615d538e9 with the bug present in
the newly added code.

In espeak, this originates from src/synthesize.cpp, with
|event->text_position| being set to |count_characters|, which is
initialized in src/readclause.cpp to |0| since speak 1.11.

Therefore, the correct value to pass to onspeaking must be:

	event->text_position