alsa: use snd_pcm_drop instead of snd_pcm_drain
If snd_pcm_drain is used when stopping audio playback, it blocks
until all the currently buffered audio is finished. This means that
pressing 'q' in the command-line application and stop in the GUI
causes the applications to hang.

Using snd_pcm_drop, the audio playback stops immediately. This
means that the applications are responsive, like when using the
simple PulseAudio APIs.