tts/engines.cpp: offset is always 0, so remove it
This was an attempt at supporting a bookmark (start at this character
offset) feature. Currently offset is always 0, so it does not
exercise the offset-supporting codepaths.

Until the way that the bookmarking feature will work is better
understood, this code is just making the logic for the speak loop
unnecessarily more complex.

There are also complications that are not solved by the current
offset code -- specifically how to correctly adjust the text length
used to calculate the total reading/recording time.