dictionary: exlude variant entries (word@number) from pronunciations
A variant entry is an entry where the word has the form:

	word@number

e.g.

	read		read@1
	read@1		/r'i:d/
	read@2		/r'Ed/

This is used when the word in question has more than one pronunciation,
allowing the different pronunciations to be specified in the dictionary
without modifying the representation.

The text to word processor can use various techniques like part of
speech analysis to determine what variant to use when given additional
data such as:

	lead		lead@1	verb
	lead		lead@2	verb past
	lead		lead@2	noun