tests/xmlcssparser: remove this experimental program
The intention behind this was to use a CSS-style specification to
describe how to parse XML-based documents. This is difficult,
because CSS does not specify how to generate RDF metadata and
structural information like that found in the WAI-ARIA spec.

As such, I am not progressing this strategy any more. It would be
useful to have another format to describe how to parse the files,
but I don't currently know what form this would take.

The approach for this new format will be something along the lines
of:

  1.  create an internal processor that works on an internal data
      model;

  2.  refactor the existing parsers to use the new processor;

  3.  define a language to describe the internal data model;

  4.  refactor the existing parsers to use the new language to
      describe the processing.