epub: initial cut at interleaving media overlay data
This takes the ePub Media Overlay events from associated SMIL
documents and interleaves them with the events from the HTML
documents.

When the text/@src property from a media overlay par group matches
the current anchor (@id) in the HTML document, the following events
are emitted:

	anchor <@id>
	media-ref <text@src> [from=audio/@clipBegin ; to=audio/@clipEnd]
	text (N)"""Lorem ipsum dolor."""

Here, the text event is the text from a subsequent event in the
HTML document.

NOTE: This is the initial logic to combine the two data streams. It
is not clear how far the media-ref event extends from the current
event stream. The ePub spec implies that the event extends the
current block scope (due to the way that XML/HTML is processed and
@id attributes behave), but this is not handled properly yet.

It also does not have any tests at the ePub level.

As the implementation progresses, the event structure may change
to meet the needs of the media overlays spec.