ffmpeg_player: don't show the '[mp3] max_analyze_duration reached' warning
On some mp3 files, the warning message:

	[mp3 0x...] max_analyze_duration reached

is displayed. This is a function of the length of the audio file,
i.e. for some files 20s is long enough while others need longer.

To simplify matters, max_analyze_duration is set to INT_MAX. While
this is overkill, it avoids complex and potentially costly
estimates that are hard to tune and get right for all cases.

Ideally, the duration of the file should not be estimated as it is
not used. The only information that is used is the codec -- that
is, the call is just used to decode the header of the file format.
There is a TODO in the ffmpeg/libav docs for
avformat_find_stream_info about allowing the user to decide what
information to fetch.

Thus, there are several options:

1.  leave the code as is;

2.  import the avformat_find_stream_info code and modify it to
    remove the information that is not needed;

3.  implement the header decoders within Cainteoir Text-to-Speech
    -- the format type is already known through the mimetype logic.