ffmpeg_player: fix low->high frequency conversion
The av_samples_get_buffer_size and avresample_convert methods have
an out_samples parameter. This is the number of samples to write
to the output buffer.

For a media player this does not typically matter, as it will use
the same frequency for the output audio as it does for the input.
This is why -- for example -- avplay.c in libav uses the frame's
nb_samples field for out_samples.

However, for Cainteoir Text-to-Speech this is not the case as the
output sample rate is determied by the selected TTS voice. For
example, the audio file being played could be at 16 kHz and the
output at 22.05 kHz.

The number of samples in a frame varies from frame to frame
(especially if the audio was encoded at a variable bit rate).
In addition to this, ffmpeg/libav do not provide a way of
calculating the number of samples in the output from a given
frame.

However, the following relationship holds:

		s_o / s_i =  f_o / f_i
	==>	s_o       = (f_o / f_i) * s_i

where s_o is the number of output samples,
      s_i is the number of input samples,
      f_o is the output frequency,
and   f_i is the input frequency.

Calculating the output samples in this way fixes the clipping of
audio when converting from a low input frequency to a high output
frequency.