ffmpeg_player: don't shadow the len variable
This fixes a bug introduced in commit 9aa978 ("ffmpeg_player: use
a single return point in resampler::resample") whereby the len
variable in the outer scope was being hidden by a len variable in
the inner scope. This resulted in len not being set and causing
garbage to be written to the audio output.