letter_to_phoneme: remove the logic from ruleset
This removes the existing logic that parsed a ruleset file and
perfomed rule matching on the text buffer to generate phonemes
according to the specified letter-to-phoneme rules.

The rationale for the removal is that the file structure and
processing logic was half-baked. As given, it would be difficult
to make the necessary improvements to the codebase to support
rulesets such as those from NRL Report 7948 or eSpeak.

The revised approach is to work on the data structure and
algorithm first -- starting with the NRL Report 7948 rules -- and
then working on the file format.