epub: add a test for an epub3 toc without entries
The test case of an epub3 nav document that does not contain any
toc entries or headings is the simplest document that can be
constructed.

At the moment this test is failing due to several bugs in the way
the epub parser is processing toc events from an epub3 html5-based
nav document.