text_reader.cpp: move the script transition logic
This change moves the logic that handled sequences of script
characters (Lo) where the script type changes from the state
transition logic to the FSM language value lookup logic.

This means that the script associated with the character is only
extracted when needed; that is only when Lo characters are
encountered.