text_reader.cpp: move newline handling to the FSM
This change moves the detection of the 'LF LF' sequence as a
paragraph block to the FSM.

The result of this is that the start position of the end-para
event is one character further along (i.e. in the middle of the
double newline sequence). Therefore, the tests have been updated
to reflect the new behaviour.