opf: remove manifest and spine from metadata tests
The OPF metadata tests included <manifest> and <spine> elements in
the metadata tests. While these are supposed to be required
elements in real OPF documents, they are not needed for the
metadata tests.

In addition to this, with the change in the OPF parser to generate
RDF Document Reference metadata for the <manifest> and <spine>
elements, these tests would fail due to the extra metadata.

Thus, these extra elements have been removed from the OPF metadata
test cases that do not need them.