tests: remove @id from the HTML5 element tests
The @id attribute can be applied to any element, not just to the
datalist and template elements it is used on in these tests. In
addition to this, the @id attribute has special behaviour.

Thus, the @id attribute will be tested more extensively in a
separate test case, keeping these test cases focused on element
processing.