xhtml: remove the old toc-handling code
The xhtml parser contained code to generate the now-removed
toc-entry events, including support for heading navigation and
partial ePub 3 table of contents. This code adds complexity to the
xhtml parser and currently does not do anything.

Thus, this code has been removed so that a cleaner implementation
can be written.