tests/events: fix the volatile HTML test failures
Given a cainteoir::buffer that looks like:

    Lorem  ipsum...
    ^     ^
    |     |
    start end

Calling cainteoir::normalize with aTrimRight=preserve resulted in
the function trimming the space at the end. This was because the
look-ahead logic to check if the next character is a whitespace
character was checking |str < l| instead of |next < l|.

This change fixes the issue and provides a test case for the bug.

I have also run the tests 100 times to see if anything else triggers
the test failures, but did not find any other failures.