phoneme: remove the afr, dpf and dpr features
These features are used to describe affricates and diphthongs.
However, this representation is wrong as it restricts the number
of these that can be represented and provides an incorrect
description of the phonemes.

The correct way to model these is to use a pair of phonemes that
describe the start and end positions of the phonemes and mark the
second phoneme as either:

1.  tied (using a tie-bar) for affricates, diphthongs or
    triphthongs; or

2.  non-syllabic for diphthongs or triphthongs.