phonemes: remove diacritics from the IPA chart
The way diacritics were displayed is complete, but generates large
HTML files, provides a clunky UI experience and deviates from the
way the official IPA chart is displayed.

This change removes the current logic for displaying diacritics.
In a future change the diacritics will be displayed in a form that
more closely resembles the official IPA chart.