ipa::phoneme -- add Kitshenbaum feature support
This is an initial mapping from the features listed in Appendix A
of Evan Kirshenbaum's ASCII-IPA encoding description to a set of
binary and ranged features.

The goals with this draft were:

  1.  Map to binary features where applicable (paired features and
      diacritics).

  2.  Map to ranged features where applicable.

  3.  Be consistent with the ASCII-IPA feature model and the IPA
      definitions/usage of those.

This forms the basis for a more complete feature set that is a
superset of (and thus backward compatible with) the Kirshenbaum
feature set.