phoneme-converter: flush output on pauses
When phoneme-converter is given the --no-pauses flag, it does not
feed the pause phoneme to the phoneme writer. This causes
pause-sensitive phonemesets (arpabet, espeak) to not produce the
correct phonemes as they are waiting for the next phoneme to
identify the transcription to use. Thus, flush is called instead
which forces the phoneme writer to clear its internal state and
output the matching phoneme.