synthesizer: iterate over samples in a duration
Given a `duration` and `sample-rate`, this iterates from time 0 to
duration as `t` in increments of `1/sample-rate`. Thus, `t` at each
step is the time at which each sample occurs.

This is the basis of a synthesizer loop.

The output is such that it can be fed into a spreadsheet program.