Rework the audio reading API.
The media playback API that was originally created was designed
specifically for playback. It sent the audio directly to an audio
object and only stopped when the audio had completed.

This has two issues:

1.  It is difficult to use in cases where the audio data needs to
    be processed (e.g. signal processing);

2.  Playback cannot be interrupted.

This commit reworks the API to a reader model. This gives control
to the caller, addressing the above issues.

NOTE: This change preserves the current behaviour on the callers.