phoneme.hpp: define a custom feature_t type
The UINT64_C constant is not defined on platforms that conform to
C99 (it is only defined under C++ if __STDC_CONSTANT_MACROS is
defined). This would affect any project that relies on the
phoneme.hpp header (or any header that includes it).

As such, the simplest solution is to define a FEATURE_C macro that
behaves the same way for a feature_t type. This also makes it
easier to update the code if/when these types change (as can be
seen in the test code).