docs: describe LPC diphone voice databases
This describes the format of the voice database (*.vdb) files that
contain Residue Excited LPC diphone voice data. It is based on the
structure of the kal voices in the flite Text-to-Speech program.

The voice database file format is designed to support additional
voice formats, but the structure of those voice formats is not
specified at this time except for the shared header data containing
information about the voice.