Rename prosody.hpp to synthesizer.hpp.
The name `prosody.hpp` was too specific to the current pho-file
model of prosody. Renaming this to `synthesizer.hpp` allows the
synthesizer API to move to this header from `engine.hpp` so that
it is co-located with the prosodic data model.

It also allows the data model to support things other than prosody
such as audio (for e.g. ePub 3 Media Overlays) and synthesizer
commands.