prosody: restructure the prosody_reader API
The prosody_reader API as previously designed was tied to data
buffers. This works with the MBROLA pho-file reader, but does not
make sense for phone to diphone conversion and converting text
events to prosody events.

Therefore, this change removes the |reset| method and moves passing
the MBROLA pho-file text buffer to the |createPhoReader| function.