context analysis: reduce the punctuation list
This change reduces the supported/mapped punctuation characters
to those that have tests associated with them. The rationale for
this is that it makes it easier to see what punctuation is tested.

In the future, these will be added back as different language text
is properly supported and tested, based on actual usage in
different texts.