context analysis: use a list for the events
The text events were stored in a queue. However, the idea is to
expose clauses as part of the API so this change to a list makes
it easier to migrate context analysis to use that API.

The reason for the API change is to allow analysis phases such as
part of speech tagging to take the whole clause into account for
better results.