clause_processor: remove the chain method
The chain method was aimed at being like the text_reader::bind
method, allowing processors to be chained together in a sequence.

This API places a burden on the implementor of the interface and
requires if checks, unlike the text_reader::bind method. To
resolve this, a helper class will be provided that does the
chaining.