phonemes: add a {rtc} feature
The {rzd} feature is used to specify rhoticized vowels, i.e. vowels
that have an R-like sound, especially the schwa vowel.

The {rtc} feature is used to specify contexts where a post-vocalic
R is pronounced in some accents as part of the syllable coda (i.e.
does not change the quality of the vowel). This is useful for
combining rhotic and non-rhotic transcriptions.