dictionary: entries with spaces must be quoted
This change means that multi-word entries need quotation marks
around them.

This is a result of a change to the way the lexer recognises text
tokens by having space characters terminate the token.

This change is needed to support more complex data formats, while
having the side-effect of making multi-word entries easier to read.