createProsodyReader: rework the text-based API
This change creates a phoneme reader that converts text events with
a clause processor into phonemes. The createProsodyReader API that
takes a text event reader and clause processor is implemented in
terms of the phoneme reader variant using this new phoneme reader.

This avoids duplication in the logic for converting phonemes to
prosody.