Use /Vʳ/ for the rhotic (r coloured) vowels.
The English transcription guide changed the way r-coloured vowels
are transcribed from /Vr/ to /Vʳ/. This is to make it easier to
process these vowels.

This change modifies tests/dict/cainteoir/words.dict to follow this
transcription guide.

The lexical set-based prosody tests are updated because they use
an accent map to transform these vowels.