mbrola/en1: support more long vowels
This adds support for several long vowels by using short vowels.

The /I/ vowel is not lengthened as the voice synthesizes it as a
clipped vowel, reflecting its tense nature.

The /@/ vowel when lengthened is a softer variant of /3:/. This is
because /@/ is designed to be used in unstressed positions.
Otherwise, the vowels sound the same. Thus, /3:/ is used for /@:/
instead of /@/.