mbrola/de*: support the /V/ and /V:/ vowels
The /V/ vowel is centralized and lowered to /6/. This matches the
phonological shift of the /V/ vowel in the modern varieties of
English where FOOT and STRUT are distinct.

The /V:/ vowel is provided for completeness, although it is not
likely to be used in practice (unless it is used for emphasis).