mbrola/de*: support /A/ and /A:/ vowels
The /A/ vowel is fronted to /a/. There does not appear to be any
documentation of this shift occuring in English accents, although
the /Q/ vowel is shifted to /a/ or /&/ in some (esp.American)
accents.

This shift produces a reasonable sounding /A/ vowel in these German
voices, especially when the long "a:" unit is used. Thus, /A/ and
/A:/ both use the "a:" unit.