mbrola/de*: support the /w/ consonant
The de1-de3 voices do not have support for the /w/ phoneme. This
adds support for /w/ in those voices by mapping it to a [u] sound.

This works because the [w] sound is in the same position as the
[u] sound, but the tongue ridge is touching the top of the mouth.

The [U] sound is not used as this would be merged with an adjacent
[U] sound, so /wU/ would sound like [U] instead of [wU]
(approximated here as [uU]).