mbrola/de*: support the /T/ and /D/ consonants
The de1-de3 voices do not have support for the /T/ and /D/ phonemes.
This adds support for these phonemes using [f] and [v] as
replacements.

This is applying th-fronting to the /T/ and /D/ phonemes, as is
common in several English accents and more accurately represents
the phonemes instead of th-alveolarization (using [s] and [z]) or
th-stopping (using [t] and [d]).

Th-alveolarization is common to German accents, but this is better
done with an accent map as the mapping here is intended to provide
the most accurate representation of the phonemes.