tests/units: add triphthong tests
This adds tests for triphthongs where the next two phonemes have
non-syllabic markers.

Triphthongs are not directly supported by the prosody model, nor
any of the voices that make use of it (i.e. the diphone-based
voices). Instead, the triphthongs are realized as a diphone pair.

That is, you have the following mapping for English triphthongs:

    /aI_^@_^/ => <aI-@>
    /aU_^@_^/ => <aU-@>