mbrola/de{1-3}: support /3:/ and /@:/
The /3:/ phoneme in English is realized as [@:] in modern English
dialects. The de1, de2 and de3 voices don't have an [3] or [3:]
phoneme, and the [@] phoneme does not support diphone pairs like
<h-@>.

Both [3:] and [@:] have been mapped to the rhotic FERN vowel as
[ER]. This gives a reasonable sounding accent, and matches the
rhoticized variants of the NEAR, CURE and SQUARE vowels used by
the German voices (with them being equivalent to the FIR, FUR and
FERN vowels).