mbrola: flush audio after a clause break
When the mbrola synthesizer is synthesizing a long phoneme
sequence, it will cut off at a certain point. This is because it
cannot synthesize a large number of phonemes in one block.

This adds special processing for the MBROLA flush unit <#>: when
the <#> unit is encountered, it will cause the current phonemes
to be read to the audio output.

The MBROLA voices have been updated so that their clause pause
phoneme entry looks like:

    unit   0 "_"
    unit 100 "#"

This will thus cause the current phonemes to be read when clause
pauses (breaks) are encountered.