Add probability distribution helpers.
The zscore class is a helper class for working with probabilistic
pitch and duration values. It takes the mean and standard-deviation
of these values and then uses a z-score to construct the actual
value at a given instance.

For the zscore class to work, it needs a way to construct the
z-score numbers. That is where the probability functions comes in.

The `zero_distribution` function provides a function that returns
0.0 for all number requests.

The `normal_distribution` function creates a random number sequence
with a mean of 0.0 and a standard distribution of 1.0.