cmudict: split the tokenizer from the parser
This splits the logic for parsing cmudict files into a lexer which
tokenizes the dictionary file, and a parser which processes the
token stream.

This simplifies the parser logic and makes it easier to add things
like error recovery. This rewrite also fixes an issue where the
last entry was not being processed.