doc2doc/mo: add a test with ids on span nodes
This structure is found in real world ePub documents containing
media overlay annotations. Currently this is not handled correctly
by the media overlay logic. The fix for this is complex and will
likely require multiple bug fixes/improvements, including to the
(X)HTML processor. Thus, the test case is documented now, but the
fix will occur later on.