doc2doc/mo: workaround span block nesting bug
This incorporates the previous media-overlays output logic which
had better handling of media references on span blocks.

The problem is that currently, the events look like:

	begin-context
	anchor
	media-ref
	text
	anchor
	media-ref
	text
	...
	end-context

This is because span tags do not emit begin/end context events. This
means that the nesting depth tracker does not work properly and
cannot correctly determine when the media-ref event ends.

The work around here -- which treats a media-ref event as ending a
previous media-ref event -- is enough to work in most cases, but
does not handle cases like:

	<p><span id=ref>Lorem</span> ipsum dolor.</p>

It also adds extra whitespace to the end of the events.

The proper fix is to include begin/end context events on all HTML
elements, allowing anchors (ids) and other constructs to be properly
tracked.