phonemes: add a {dcz} (diacritize) feature
Various diacritics use IPA phoneme symbols (aspirated, velarized,
etc.) to denote some specialized aspect of the phoneme. These are
currently supported in the Cainteoir Engine by using specific
features for these diacritics. The problem with this is that the
following annotations complicate the model:

  1.  pre-aspirated consonants (e.g. /^{h}p/);

  2.  double aspirated consonants (e.g. /p^{hh}/);

  3.  additional superscript forms (fricative release, breathy voice,
      glottalized, compressed, etc.);

  4.  alternate usage (e.g. for /o^{w}/ the /^{w}/ is off-glide, not
      labialized);

  5.  using superscript forms for diphthongs.

The approach taken here is to use {dcz} to denote a superscript form
and leave the interpretation of that up to the other layers. Thus,
/p^{h}/ would be modelled as /{vls,blb,stp}{vls,glt,frc,dcz}/.

This change just adds the {dcz} feature model.