L2P: recurse into next_match when ignoring elision
When in the ignore elision state (i.e. when finding the next phoneme
to match against), recursively call next_match until a phoneme is
found.

This ensures that as long as there are characters to match (i.e. the
elided phoneme is not at the end of the word), there will always be
a next phoneme.